requirements.txt
Beaker==1.6.4
Chameleon==2.11
Mako==0.7.3
MarkupSafe==0.15
PasteDeploy==1.5.0
Pygments==1.6
SQLAlchemy==0.8.0
WebOb==1.2.3
WebTest==2.0.4
Whoosh==2.4.1
beautifulsoup4==4.1.3
cov-core==1.7
coverage==3.6
creme-fraiche==0.0
mock==1.0.1
psycopg2==2.4.6
py==1.4.13
pyramid==1.4
pyramid-beaker==0.7
pyramid-debugtoolbar==1.0.4
pyramid-tm==0.7
pytest==2.3.4
pytest-cov==1.6
python-ldap==2.4.10
repoze.lru==0.6
six==1.3.0
transaction==1.4.1
translationstring==1.1
venusian==1.0a7
waitress==0.8.2
wsgiref==0.1.2
zope.deprecation==4.0.2
zope.interface==4.0.5
zope.sqlalchemy==0.7.2
